requirements.txt
fastapi
uvicorn
pydantic
python-multipart
fastmcp
python-dotenv
httpx
pydantic